app/requirements.txt
mysql-connector-python
unittest